extracting widgets
